adjusts git ignore and adds prettier
I needed to install prettier within my project to get it to function
properly with angular 17. Hopefully i can find a way around this in the
future as it's a little sloppy. I also did some research and it seems
like NPM wants you to export package-lock to ensure deterministic
install so I'm going to start doing that.