reimplements angular material themes
I decided for the time being it would just be easier to go back to angular material themes to get everything to be colored more or less properly. I still am thinking it will just be easier in the long run to build all the components myself, but this works fairly well for now.