refactor game components
The p5/components directory was starting to feel a little bloated. I also just finally broke down and accepted it's simpler to keep all my CSS for one game in a single file. If there are any overrides they're small enough to just put in the styles array in the .ts file.