another layout for demon-entry div
The only way I'm going to get the text to fit in the SVG I made the same way it does in the game is if I'm willing to let the text get unreadably small. So I decided to keep the SVG but not really try to align with it.