add support for 4+ skill special fusions
I split up demon_getBuildRecipes into to functions depending on if our result needs to inherit more than 4 skills or not. Then I added a second optional set of properties to BuildRecipe to accomadate the cases when there are two concurrent recipes necessary for a build. Then I just brute forced the HTML to display two recipe if these optional properties exit. Not really happy with that, but I don't know where else to do this other than HTML if I wanna style it.